JPA enum 타입 생성 및 컨버터 추가 및 연동 완료

+++ src/main/java/net/softbell/bsh/iot/dto/bshp/v1/ItemInfoV1DTO.java
@@ -17,7 +17,7 @@ public class ItemInfoV1DTO
 {
 	private Byte controlMode;
 	private Byte pinId;
-	private Byte pinMode;
-	private Byte pinType;
+	private Integer pinMode;
+	private Integer pinType;
 	private String pinName;
 }
